backend/submission/Mohamed_Abdallah/requirements.txt
Django==5.1.5
djangorestframework==3.15.1
drf-nested-routers==0.93.4
